+++ apps/backoffice-v2/src/pages/MerchantMonitoring/components/MerchantMonitoringTable/MerchantMonitoringTable.tsx
@@ -1,0 +1,24 @@
+import React, { FunctionComponent } from 'react';
+import { TBusinessReports } from '@/domains/business-reports/fetchers';
+import { DataTable } from '@/common/components/organisms/DataTable/DataTable';
+import { columns } from '@/pages/MerchantMonitoring/components/MerchantMonitoringTable/columns';
+import { useMerchantMonitoringTableLogic } from '@/pages/MerchantMonitoring/components/MerchantMonitoringTable/hooks/useMerchantMonitoringTableLogic/useMerchantMonitoringTableLogic';
+
+export const MerchantMonitoringTable: FunctionComponent<{
+  data: TBusinessReports;
+}> = ({ data }) => {
+  const { Cell } = useMerchantMonitoringTableLogic();
+
+  return (
+    <DataTable
+      data={data}
+      columns={columns}
+      CellContentWrapper={Cell}
+      sortByField={`createdAt`}
+      options={{
+        enableSorting: true,
+      }}
+      props={{ scroll: { className: 'h-full' }, cell: { className: '!p-0' } }}
+    />
+  );
+};