+++ services/workflows-service/prisma/migrations/20240526072522_business_report_status/migration.sql
@@ -1,0 +1,11 @@
+/*
+  Warnings:
+
+  - Added the required column `status` to the `BusinessReport` table without a default value. This is not possible if the table is not empty.
+
+*/
+-- CreateEnum
+CREATE TYPE "BusinessReportStatus" AS ENUM ('in_progress', 'completed');
+
+-- AlterTable
+ALTER TABLE "BusinessReport" ADD COLUMN     "status" "BusinessReportStatus" NOT NULL;