+++ packages/workflow-core/CHANGELOG.md
@@ -1,5 +1,33 @@
 # @ballerine/workflow-core
 
+## 0.6.14
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.8
+
+## 0.6.13
+
+### Patch Changes
+
+- version bump
+
+  s Please enter a summary for your changes.
+
+## 0.6.12
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.9.7
+
+## 0.6.11
+
+### Patch Changes
+
+- Version bump
+
 ## 0.6.10
 
 ### Patch Changes